requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Jsonpify==1.5.0
Flask-JWT
flask-marshmallow==0.7.0
flask-mongoengine==1.0.0
Flask-OAuthlib==0.9.6
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
flask-restplus==0.11.0
Flask-SQLAlchemy==2.4.3
flask-swagger==0.2.14
flask-swagger-ui==3.36.0
Flask-WTF==1.0.1
Jinja2==3.0.2
jsonschema==4.5.1
MarkupSafe==2.0.1
marshmallow==2.21.0
marshmallow-sqlalchemy==0.12.1
mongoengine==0.24.1
platformdirs==2.4.0
PyJWT
pymongo==3.12.1
python-dotenv==0.20.0
pytz==2020.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.1.0
webargs==1.10.0
Werkzeug==2.0.2
bcrypt==3.2.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.1